Project enhanced space objects at subpixel positions

+++ include/wc1sdl.h
@@ -155,7 +155,7 @@ void Wc1SdlBeginSpaceFrame(
 void Wc1SdlCompleteSpaceFrame(void);
 void Wc1SdlCancelSpaceFrame(void);
 int Wc1SdlRecordSpaceSprite(
-    const struct Viewport *viewport, short x, short y,
+    const struct Viewport *viewport, float x, float y,
     unsigned char *shape, short frame, short angle, short scale,
     short flip);
 int Wc1SdlSetCursorPosition(int x, int y);
